+++ src/neon/loader/TextureLoader.cpp
@@ -85,7 +85,7 @@ namespace neon {
             auto optional = getFile(std::filesystem::path(data.get<std::string>()), context);
             if (optional.has_value()) {
                 datas.push_back(optional.value().getData());
-                sizes.push_back(optional.value().getSize());
+                sizes.push_back(static_cast<uint32_t>(optional.value().getSize()));
                 files.push_back(std::move(optional.value()));
             }
         }
